src/pyproject.toml
[tool.poetry]
name = "3DPoseEvaluator"
version = "2.0.0"
description = "Program to evaluate NIST-provided human locomotion video sequences with 3D pose estimation algorithms: GAST-NET, VIBE, and Blazepose"
authors = ["Amaan Rahman <amaan.rahman.usa@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
matplotlib = "^3.5.1"
numpy = "^1.22.0"
scipy = "^1.7.3"
pandas = "^1.3.5"
seaborn = "^0.11.2"
moviepy = "^1.0.3"
resampy = "^0.2.2"
pyquaternion = "^0.9.9"
opencv-python = "^4.5.5"
joblib = "^1.1.0"
argparse = "^1.4.0"
Cython = "^0.29.26"
setuptools = "^60.5.0"
ffmpeg = "^1.4"
numba = "^0.55.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

